+++ src/cpu/fetch.rs
@@ -4,7 +4,7 @@ use super::decode::Decode;
 
 pub fn fetch(cpu: &mut super::CPU) -> Result<Box<dyn Decode>, (Option<u32>, TrapCause, String)> {
     dbg_hex!(cpu.pc);
-    let index_pc: u32 = cpu.trans_addr(TransFor::Fetch, cpu.pc as i32)?;
+    let index_pc: u32 = cpu.trans_addr(TransFor::Fetch, cpu.pc)?;
     let is_cinst: bool = cpu.bus.raw_byte(index_pc) & 0x3 != 0x3;
 
     if is_cinst {
@@ -13,7 +13,7 @@ pub fn fetch(cpu: &mut super::CPU) -> Result<Box<dyn Decode>, (Option<u32>, Trap
             (cpu.bus.raw_byte(index_pc) as u16);
         Ok(Box::new(new_inst))
     } else {
-        let index_pc2: u32 = cpu.trans_addr(TransFor::Fetch, (cpu.pc + 2) as i32)?;
+        let index_pc2: u32 = cpu.trans_addr(TransFor::Fetch, cpu.pc + 2)?;
         let new_inst: u32 =
             (cpu.bus.raw_byte(index_pc2 + 1) as u32) << 24 |
             (cpu.bus.raw_byte(index_pc2) as u32) << 16 |
